Remove --cloud option from integration tests
